Split resolver metadata helpers
